add has_many :invitations association to Meeting model; change minutes_duration to float in db

+++ db/migrate/20210914170821_create_meetings.rb
@@ -2,7 +2,7 @@ class CreateMeetings < ActiveRecord::Migration[6.1]
   def change
     create_table :meetings do |t|
       t.string :title, default: "New Meeting"
-      t.integer :minutes_duration, default: 30
+      t.float :minutes_duration, default: 30
       t.datetime :start_time, null: false, default: DateTime.now
 
       t.timestamps
